// crud-gen/src/ag-grid.interface.ts
export interface IBaseFilterModel {
  filterType: FilterType;
  field: string;
}

export interface ISimpleFilterModel extends IBaseFilterModel {
  type: GeneralFilters;

  filter?: string | number;
}

export interface ITextFilterModel extends ISimpleFilterModel {
  // always 'text' for text filter
  filterType: FilterType.TEXT;

  // one of the filter options, e.g. 'equals'
  type: GeneralFilters;

  // the text value associated with the filter.
  // it's optional as custom filters may not
  // have a text value
  filter?: string;
}

export interface ISetFilterModel extends IBaseFilterModel {
  filterType: FilterType.SET;

  values: (string | number)[];
}

export interface INumberFilterModel extends ISimpleFilterModel {
  // always 'number' for number filter
  filterType: FilterType.NUMBER;

  // one of the filter options, e.g. 'equals'
  type: GeneralFilters;

  // the number value(s) associated with the filter.
  // custom filters can have no values (hence both are optional).
  // range filter has two values (from and to).
  filter?: number;
  filterTo?: number;
}

export interface DateFilterModel extends ISimpleFilterModel {
  // always 'date' for date filter
  filterType: FilterType.DATE;

  // one of the filter options, e.g. 'equals'
  type: GeneralFilters;

  // the date value(s) associated with the filter.
  // the type is string and format is always YYYY-MM-DD e.g. 2019-05-24
  // custom filters can have no values (hence both are optional).
  // range filter has two values (from and to).
  dateFrom?: string;
  dateTo?: string;
}

export type GenericFilterModel =
  | ISimpleFilterModel
  | ITextFilterModel
  | INumberFilterModel;

export type FilterModel =
  | GenericFilterModel
  | DateFilterModel
  | ISetFilterModel;

export interface ICombinedSimpleModel {
  // the filter type: date, number or text
  filterType: FilterType;

  operator: Operators;

  // two instances of the filter model
  condition1: FilterModel;
  condition2: FilterModel;
}
// ...
export interface IFilterExpressionsProperty {
  [FilterType.TEXT]?: ITextFilterModel;
  [FilterType.NUMBER]?: INumberFilterModel;
  [FilterType.DATE]?: DateFilterModel;
  [FilterType.SET]?: ISetFilterModel;
  [FilterType.MULTI]?: never;
}

export type FilterExpressionType =
  | ITextFilter
  | INumberFilter
  | IDateFilter
  | ISetFilter;

export interface FilterInput {
  operator?: Operators;
  expressions?: IFilterExpressionsProperty[];
  childExpressions?: FilterInput[];
}